In order to reproduce the original Koning-Delaroche optical potential, a sign needs to be changed in one of two places in 10.1103/PhysRevC.107.014602:
  Option 1: The '±' in front of w^{n,p}_{1,A} in Eq. (A11) from the main
text should be replaced by just '+' (note also the missing factor of A
which appears to be left out)
  Option 2: A minus sign should be added in front of the tabulated value
of "W1_A_n" given in the JSON table printed in supplement_parameters.pdf.

Previous work on the optical model JSON configuration assumed Option 2.
This commit flips the interpretation to prefer Option 1. The factor of A
mentioned in Option 1 was always included in the MARLEY implementation.